chore: land the in-flight work this checkout holds
The owner asked for the first green push of what exists, and what exists
includes 45 files of other sessions' work sitting uncommitted in this
shared checkout. Committing it is what makes a full verification judge
the TREE rather than a HEAD the tree has moved past, which is the
mistake this session made twice before he corrected it.
The work spans BGP config and plugins, IKE engine, the FIB and iface
plugins, the CLI client and completer, managed, L2TP, AAA, core
configvalue and diagnostic, six interoplab checkers, and four le
packages. It compiles: go vet over the whole tree reports only a
pre-existing unsafe.Pointer heuristic in internal/core/textbuf, the
zero-allocation buffer where that use is deliberate and where nothing is
modified.
Tested before landing rather than after: 98 packages green over every
changed area. Four are red and none is caused by this work.
filter_modify, route_refresh/handler and cli/client carry no modified
file at all. bgp/config is red through peer_keywords_test.go and
loader_test.go, both unmodified, reporting that the reservedPeerNames
entries update and raw name no keyword a bgp peer container declares;
ze-raw-cmd.yang and ze-update-cmd.yang are untouched since 2026-09-05,
and plan/journal/helper-bypassed-by-an-open-coded-copy.md records the
selector migration moving those verbs under request peer. The four are
red at HEAD and stay red at HEAD.
One RFC-tagged test changes and test/rfc-changed/c9bdcd62.md records why
its claim does not: TestLocRIBRouteMonitoringUsesFourByteASNs passes its
entry by pointer, following a producer signature, and still asserts the
same four-byte ASN encoding.
continue.md lands with it. It is the handover from the session that
closed three VyOS-derived subscriber specs, and it names what is still
owed: two approved specs not started, four decisions waiting on the
owner, and a closed spec rewritten to disk after its own removal commit,
which never-destroy-work correctly stopped that session from deleting
unasked.

+++ internal/component/ike/engine/rekey.go
@@ -540,7 +540,7 @@ func respondChildRekey(sa *SA, inner []wire.PayloadEntry, old *ChildSA, msgID ui
 	// exchange is not keyed. RFC 7296 Section 3.10.1 gives the payload two octets of data.
 	// respondIKERekey answers an IKE SA rekey the same way.
 	if group != dhGroupNone && (reqKE == nil || reqKE.DHGroup != uint16(group)) {
-		got := uint16(wire.DHGroupNone)
+		got := wire.DHGroupNone
 		if reqKE != nil {
 			got = reqKE.DHGroup
 		}
